+++ src/com/intellij/idea/plugin/hybris/polyglotQuery/actions/PolyglotQueryExecuteQueryAction.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
- * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -19,29 +19,14 @@ package com.intellij.idea.plugin.hybris.polyglotQuery.actions
 
 import com.intellij.idea.plugin.hybris.actions.AbstractExecuteAction
 import com.intellij.idea.plugin.hybris.common.HybrisConstants
-import com.intellij.idea.plugin.hybris.common.utils.HybrisI18NBundleUtils
+import com.intellij.idea.plugin.hybris.common.utils.HybrisI18NBundleUtils.message
 import com.intellij.idea.plugin.hybris.common.utils.HybrisIcons
 import com.intellij.idea.plugin.hybris.polyglotQuery.file.PolyglotQueryFileType
-import com.intellij.openapi.actionSystem.AnActionEvent
-import com.intellij.openapi.actionSystem.CommonDataKeys
 
 class PolyglotQueryExecuteQueryAction : AbstractExecuteAction(
     PolyglotQueryFileType.defaultExtension,
-    HybrisConstants.CONSOLE_TITLE_POLYGLOT_QUERY
-) {
-
-    init {
-        with(templatePresentation) {
-            text = HybrisI18NBundleUtils.message("hybris.pgq.actions.execute_query")
-            description = HybrisI18NBundleUtils.message("hybris.pgq.actions.execute_query.description")
-            icon = HybrisIcons.Console.Actions.EXECUTE
-        }
-    }
-
-    override fun update(e: AnActionEvent) {
-        super.update(e)
-        val file = e.dataContext.getData(CommonDataKeys.VIRTUAL_FILE)
-        val enabled = file != null && file.name.endsWith(".$extension")
-        e.presentation.isEnabledAndVisible = enabled
-    }
-}
+    HybrisConstants.CONSOLE_TITLE_POLYGLOT_QUERY,
+    message("hybris.pgq.actions.execute_query"),
+    message("hybris.pgq.actions.execute_query.description"),
+    HybrisIcons.Console.Actions.EXECUTE
+)